Move CLISearchProgress to correct package
