+++ MultiSubDownloader/src/main/java/org/lodder/subtools/multisubdownloader/CLI.java
@@ -9,9 +9,9 @@ import org.lodder.subtools.multisubdownloader.actions.DownloadAction;
 import org.lodder.subtools.multisubdownloader.actions.FileListAction;
 import org.lodder.subtools.multisubdownloader.actions.SubtitleSelectionAction;
 import org.lodder.subtools.multisubdownloader.cli.progress.CLIFileindexerProgress;
+import org.lodder.subtools.multisubdownloader.cli.progress.CLISearchProgress;
 import org.lodder.subtools.multisubdownloader.framework.Container;
 import org.lodder.subtools.multisubdownloader.gui.actions.search.CliSearchAction;
-import org.lodder.subtools.multisubdownloader.gui.dialog.progress.search.CLISearchProgress;
 import org.lodder.subtools.multisubdownloader.lib.Info;
 import org.lodder.subtools.multisubdownloader.lib.ReleaseFactory;
 import org.lodder.subtools.multisubdownloader.lib.SubtitleSelectionCLI;
